Fix highlight matching and add search highlight tests
